added the necessary assets, folders and files, moved my app.jsx logic, state and component rendering into my new Home.jsx file, and made my App.jsx file the main router file.
