+++ src/pages/upload/Upload.jsx
@@ -1,0 +1,15 @@
+import Header from './../../components/Header/Header';
+import UploadHero from './../../components/UploadHero/UploadHero';
+import UploadForm from '../../components/UploadForm/UploadForm';
+
+const Upload = () => {
+    return (
+        <>
+            <Header/>
+            <UploadHero/>
+            <UploadForm/>
+        </>
+    )
+}
+
+export default Upload;
