+++ src/pages/upload/Upload.scss
@@ -1,0 +1,1 @@
+@import "./../../styles/partials/global.scss";
